move notification type to redux and omit some imports
