+++ tensorboard/webapp/notification_center/_views/notification_center_component.ts
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 import {Component, Input} from '@angular/core';
-import {Notification} from '../types';
+import {Notification} from '../_redux/notification_center_types';
 
 @Component({
   selector: 'notification-center-component',
